requirements.txt
#
# Pinned minimal stack for LangGraph project
#
langchain>=0.3.7
langgraph>=0.2.34
langchain-openai>=0.2.3
langchain-core>=0.3.13
pydantic>=2.9.2
fastapi>=0.115.5
uvicorn>=0.32.0
python-dotenv>=1.0.1
duckduckgo-search>=6.3.6
httpx>=0.27.2
sqlalchemy>=2.0.36
tavily>=0.1.14
langgraph-checkpoint-sqlite>=0.1.0
